move island around just the select instead of the whole page
