+++ dotcom-rendering/src/components/FootballTablesCompetitionSelect.importable.tsx
@@ -1,0 +1,27 @@
+import type { Region } from '../footballMatches';
+import { FootballCompetitionSelect } from './FootballCompetitionSelect';
+
+const goToCompetitionSpecificPage =
+	(guardianBaseUrl: string) => (path: string) => {
+		const url = `${guardianBaseUrl}${path}`;
+		window.location.assign(url);
+	};
+
+type Props = {
+	regions: Region[];
+	pageId: string;
+	guardianBaseUrl: string;
+};
+
+export const FootballTablesCompetitionSelect = ({
+	regions,
+	pageId,
+	guardianBaseUrl,
+}: Props) => (
+	<FootballCompetitionSelect
+		regions={regions}
+		pageId={pageId}
+		onChange={goToCompetitionSpecificPage(guardianBaseUrl)}
+		kind="Tables"
+	/>
+);
