+++ dotcom-rendering/src/components/FootballTablesPage.stories.tsx
@@ -1,5 +1,4 @@
 import type { Meta, StoryObj } from '@storybook/react/*';
-import { fn } from '@storybook/test';
 import { regions } from '../../fixtures/manual/footballData';
 import { FootballTableList as TableListDefault } from './FootballTableList.stories';
 import { FootballTablesPage as FootballTablesPageComponent } from './FootballTablesPage';
@@ -16,7 +15,6 @@ type Story = StoryObj<typeof meta>;
 export const FootballTablesPage = {
 	args: {
 		regions,
-		goToCompetitionSpecificPage: fn(),
 		pageId: 'football/tables',
 		competitions: TableListDefault.args.competitions,
 		renderAds: true,
